Extract `VisitRef`.
